src/ContosoSuitesDashboard/requirements.txt
pandas
numpy
scikit-learn
streamlit==1.29.0
streamlit-extras==0.3.6
streamlit_js_eval==0.1.5
openai==1.12.0
azure-cognitiveservices-speech>=1.33.0
azure-ai-textanalytics>=5.3.0